RF: move setup_ teardown_ to the top level __init__.py
