+++ mvpa2/tests/__init__.py
@@ -9,11 +9,7 @@
 """Unit test interface for PyMVPA"""
 from __future__ import print_function
 
-from builtins import str
-import sys
 import unittest
-import numpy as np
-from mvpa2 import _random_seed, cfg, wtf
 from mvpa2.base import externals, warning
 
 # # init to make tests into a package
@@ -21,75 +17,6 @@ from mvpa2.base import externals, warning
 # from numpy.testing import Tester
 # test = Tester().test
 # del Tester
-
-_sys_settings = {}
-
-def _get_verbosity(verbosity):
-    if verbosity is None:
-        return int(cfg.get('tests', 'verbosity', default=1))
-    return verbosity
-
-def setup_module(module, verbosity=None):
-    "set up test fixtures for testing"
-
-    if __debug__:
-        from mvpa2.base import debug
-        # Lets add some targets which provide additional testing
-        debug.active += ['CHECK_.*']
-
-    verbosity = _get_verbosity(verbosity)
-
-    # provide people with a hint about the warnings that might show up in a
-    # second
-    if verbosity:
-        print("T: MVPA_SEED=%s" % _random_seed)
-        if verbosity > 1:
-            print('T: Testing for availability of external software packages.')
-
-    # fully test of externals
-    verbosity_dependencies = max(0, verbosity - 1)
-    if verbosity_dependencies:
-        externals.check_all_dependencies(verbosity=verbosity_dependencies)
-    elif __debug__ and verbosity: # pragma: no cover
-        print('T: Skipping testing of all dependencies since verbosity '
-              '(MVPA_TESTS_VERBOSITY) is too low')
-
-    # So we could see all warnings about missing dependencies
-    _sys_settings['maxcount'] = warning.maxcount
-    warning.maxcount = 1000
-
-    if verbosity < 3:
-        # no MVPA warnings during whole testsuite (but restore handlers later on)
-        _sys_settings['handlers'] = warning.handlers
-        warning.handlers = []
-
-        # No python warnings (like ctypes version for slmr)
-        import warnings
-        warnings.simplefilter('ignore')
-
-    if verbosity < 4:
-        # No NumPy
-        _sys_settings['np_errsettings'] = np.geterr()
-        np.seterr(**dict([(x, 'ignore') for x in _sys_settings['np_errsettings']]))
-
-
-def teardown_module(module, verbosity=None):
-    "tear down test fixtures"
-    verbosity = _get_verbosity(verbosity)
-
-    # restore warning handlers
-    warning.maxcount = _sys_settings['maxcount']
-
-    if verbosity < 3:
-        # restore warning handlers
-        warning.handlers = _sys_settings['handlers']
-
-    if verbosity < 4:
-        # restore numpy settings
-        np.seterr(**_sys_settings['np_errsettings'])
-
-    if cfg.getboolean('tests', 'wtf', default='no'):
-        sys.stderr.write(str(wtf()))
 
 
 def collect_unit_tests(verbosity=1):
@@ -281,6 +208,7 @@ def run_tests_using_nose(limit=None, verbosity=1, exit_=False):
     TODO: just switch to using numpy.testing framework, for that
           unittests need to be cleaned and unified first
     """
+    raise NotImplementedError("Likely would be sacrificed")
     nosetests = collect_nose_tests(verbosity=verbosity)
     verbosity = _get_verbosity(verbosity)
 
@@ -290,7 +218,6 @@ def run_tests_using_nose(limit=None, verbosity=1, exit_=False):
         return
 
     from nose import main
-    import nose
     import nose.config
 
     tests = collect_unit_tests(verbosity=verbosity) + nosetests
@@ -338,7 +265,7 @@ def run(limit=None, verbosity=None, exit_=False):
     exit_ : bool, optional
       Either to exit with an error code upon the completion.
     """
-
+    from mvpa2 import setup_module, teardown_module, _random_seed
     setup_module(None, verbosity)
 
     try:
